Make puzzles ID unsigned long long

+++ a_star.cpp
@@ -9,7 +9,7 @@
 */
 
 Solution solve_a_star(vector<int> user_input, HeuristicCalculator h){
-    int puzzle_counter = 1;
+    unsigned long long puzzle_counter = 1;
     int num_expanded_nodes = 0;
     priority_queue<Puzzle*, vector<Puzzle*>, AStarComparator>* open = new priority_queue<Puzzle*, vector<Puzzle*>, AStarComparator>();
     unordered_set<unsigned long long>* closed = new unordered_set<unsigned long long>();
